Refs Uni-Lab-OS/Uni-Lab-Core#166

+++ szlab_poly_studio/workflows/szlab_mixer.py
@@ -9,7 +9,7 @@ szlab_mixer_pump: SzlabMixerPumpDevice = device("szlab_mixer_pump")
     workflow_uuid="33930b26-2780-555f-8668-e56462846716",
     displayname="S06 加液调试",
 )
-def szlab_mixer_workflow(*, pump: int = 1, volume: float = 8.0) -> None:
+def szlab_mixer_workflow(*, pump: int = 1, volume: int = 8) -> None:
     # unilab:node_uuid=b6568882-960a-539f-b70a-63194267b086
     addition = szlab_mixer_pump.run_solvent_addition(  # noqa: F841
         pump=pump,
